third_party/WebKit/LayoutTests/editing/selection/unrendered-space.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div contenteditable id="div">foo\n<div>bar</div></div>',
    selection => {
        var div = selection.document.getElementById("div");
        var text = div.firstChild;
        selection.collapse(text, text.length);
    },
    '<div contenteditable id="div">foo\n|<div>bar</div></div>'),
    'Visible position creation in unrenderer space after renderer text.');
</script>
